% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_adm_data}
\alias{clean_adm_data}
\title{Apply standardized data cleaning opperations}
\usage{
clean_adm_data(df)
}
\arguments{
\item{df}{a data frame to clean}
}
\value{
a cleaned data frame
}
\description{
Apply standardized data cleaning opperations
}
\examples{
\dontrun{clean_adm_data(df)}
}
