% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ice_data.R
\name{clean_ice_data}
\alias{clean_ice_data}
\title{Clean ICE data using standard rfcip patterns}
\usage{
clean_ice_data(data)
}
\arguments{
\item{data}{Data frame to clean}
}
\value{
Cleaned data frame
}
\description{
Clean ICE data using standard rfcip patterns
}
\keyword{internal}
