% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_ice_description}
\alias{extract_ice_description}
\title{Extract human-readable description from ICE filename}
\usage{
extract_ice_description(filename)
}
\arguments{
\item{filename}{Character filename to parse}
}
\value{
Character description
}
\description{
Extract human-readable description from ICE filename
}
\keyword{internal}
