% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_data.R
\name{get_col_data}
\alias{get_col_data}
\title{Builds a cause of loss dataset using data from the specified years}
\usage{
get_col_data(
  years = c(as.numeric(format(Sys.Date(), "\%Y")):as.numeric(format(Sys.Date(), "\%Y")) -
    4)
)
}
\arguments{
\item{years}{a single numeric value or vector of values specifying the commodity years that should be used to construct the cause of loss data set}
}
\value{
returns a data frame with cause of loss data corresponding to the specified commodity years
}
\description{
Builds a cause of loss dataset using data from the specified years
}
\examples{
\dontrun{
get_col_data(years = 2020)
get_col_data(years = 2019:2021)
}
}
