% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ice_data.R
\name{get_ice_data_single_year}
\alias{get_ice_data_single_year}
\title{Get ICE data for a single year}
\usage{
get_ice_data_single_year(year, dataset, ice_url, show_progress, force)
}
\arguments{
\item{year}{Single numeric year}

\item{dataset}{Character dataset filter}

\item{ice_url}{Character URL}

\item{show_progress}{Logical for progress display}

\item{force}{Logical for cache bypass}
}
\value{
Data frame with ICE data
}
\description{
Get ICE data for a single year
}
\keyword{internal}
