% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ice_data.R
\name{get_ice_metadata}
\alias{get_ice_metadata}
\title{Get ICE metadata for specified years}
\usage{
get_ice_metadata(years, ice_url)
}
\arguments{
\item{years}{Numeric vector of years}

\item{ice_url}{Character URL for ICE repository}
}
\value{
Data frame with file metadata
}
\description{
Get ICE metadata for specified years
}
\keyword{internal}
