% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_sobtpu_data}
\alias{get_sobtpu_data}
\title{Get summary of business data by type, practice, and unit structure.}
\usage{
get_sobtpu_data(
  year = NULL,
  crop = NULL,
  insurance_plan = NULL,
  state = NULL,
  county = NULL,
  fips = NULL,
  cov_lvl = NULL
)
}
\arguments{
\item{year}{a numeric value (either single value or a vector of values) that indicate the crop year (ex: \code{2024} or \code{c(2022,2023,2024)})}

\item{crop}{can be either a character string indicating a crop (i.e. \code{corn}) or a numeric value indicating the crop code (i.e. \code{41}). Inputting a vector with multiple values will return data for multiple crops. To get a data frame containing all the available crops and crop codes use \code{get_crop_codes()}.}

\item{insurance_plan}{can be either a character string indicating an insurance plan (ex: \code{yp} and \verb{yield protection} are both valid) or a numeric value indicating the insurance plan code (i.e. \code{1}). Inputting a vector with multiple values will return data for multiple insurance plans. To get a data frame containing all the available insurance plans and insurance plan codes use \code{get_insurance_plan_codes()}.}

\item{state}{can be a character string indicating the state abbreviation or state name. Numeric values corresponding to state FIPS codes can also be supplied.}

\item{county}{either a character string with a county name or 5-digit FIPS code corresponding to a county. when the county is specified using the name of the county, the state parameter must also be specified.}

\item{fips}{a numeric value corresponding to a 5-digit FIPS code of a U.S. county.}

\item{cov_lvl}{a numeric value indicating the coverage level. Valid coverage levels are \code{c(0.5,0.55,0.6,0.65,0.7,0.75,0.8,0.85,0.9,0.95)}}
}
\value{
A tibble containing the summary of business data by type, practice, and unit structure.
}
\description{
This function is a helper function for \code{get_sob_data()} that gets activated when a user sets \code{sob_version = "sobtpu"} in \code{get_sob_data()}.
All of the arguments in \code{get_sobtpu_data} are inherited from \code{get_sob_data()}. The function can be used as a standalone function however,
but is not exported, meaning it must be called via \code{rfcip:::get_sobtpu_data()}.
}
