% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_livestock_adm_data.R
\docType{data}
\name{LIVESTOCK_ADM_DATASETS}
\alias{LIVESTOCK_ADM_DATASETS}
\title{Livestock ADM dataset code mapping}
\format{
An object of class \code{list} of length 7.
}
\usage{
LIVESTOCK_ADM_DATASETS
}
\description{
Maps user-friendly dataset names to the codes/patterns used in
the FTP directory filenames.
}
\keyword{internal}
