% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_col_zip}
\alias{process_col_zip}
\title{Process cause of loss ZIP file}
\usage{
process_col_zip(zip_file)
}
\arguments{
\item{zip_file}{Character. Path to cached ZIP file.}
}
\value{
Data frame with processed cause of loss data.
}
\description{
Extracts and processes data from a cached cause of loss ZIP file.
}
\keyword{internal}
