% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcip-package.R
\docType{package}
\name{rfcip-package}
\alias{rfcip}
\alias{rfcip-package}
\title{rfcip: What the Package Does (One Line, Title Case)}
\description{
This package is a wrapper for RMA's summary of business app. The package helps users get data on the Federal Crop Insurance Program in an easy to use format.
}
\author{
\strong{Maintainer}: Dylan Turner \email{dylan.turner3@outlook.com} (\href{https://orcid.org/0000-0002-0915-7384}{ORCID})

}
\keyword{internal}
