% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_insurance_plan_codes.R
\name{get_insurance_plan_codes}
\alias{get_insurance_plan_codes}
\title{Lookup insurance plan codes for FCIP insurance plans}
\source{
Data is downloaded directly from RMA's summary of business app: \url{https://public-rma.fpac.usda.gov/apps/SummaryOfBusiness/ReportGenerator}
}
\usage{
get_insurance_plan_codes(
  year = as.numeric(format(Sys.Date(), "\%Y")),
  plan = NULL
)
}
\arguments{
\item{year}{A single numeric value, or vector of numeric values, indicating what years of the summary of business should be used to get crop names and crop codes. Defaults to the current year.}

\item{plan}{can be either a character string indicating an insurance plan (ex: \code{yp} and \verb{yield protection} are both valid) or a numeric value indicating the insurance plan code (i.e. \code{1}). Inputting nothing for the \code{plan} argument will return codes for all insurance plans in the specified year(s).}
}
\value{
Returns a tibble containing the relevant commodity year, insurance plan codes, insurance plan name, and insurance plan abbreviation
}
\description{
Lookup insurance plan codes for FCIP insurance plans
}
\examples{
\dontrun{
get_insurance_plan_codes()
get_insurance_plan_codes(year = 2023, plan = "yp")
get_insurance_plan_codes(year = 2023, plan = 1)
get_insurance_plan_codes(year = 2023, plan = "yield protection")
get_insurance_plan_codes(year = 2018:2022, plan = c("yp", "rp"))
}
}
