% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price_data.R
\name{get_price_data}
\alias{get_price_data}
\title{Download RMA projected and harvest prices used for revenue protection insurance plans.}
\source{
Data is downloaded directly from RMA's price discovery app: \url{https://public-rma.fpac.usda.gov/apps/PriceDiscovery/GetPrices/ManyPrices}
}
\usage{
get_price_data(year = NULL, crop = NULL, state = NULL, force = FALSE)
}
\arguments{
\item{year}{a numeric value (either single value or a vector of values) that indicate the crop year (ex: \code{2024} or \code{c(2022,2023,2024)})}

\item{crop}{can be either a character string indicating a crop (i.e. \code{corn}) or a numeric value indicating the crop code (i.e. \code{41}).
Inputting a vector with multiple values will return data for multiple crops.
To look up crop codes, use the \code{get_crop_codes()} function.}

\item{state}{either a numeric value representing the state fips code or a character string that can either be the full state name or two-digit state abbreviation.}

\item{force}{logical (default FALSE). If TRUE, attempts to download fresh data regardless of cache, but falls back to cached data on failure with a warning}
}
\value{
Returns a tibble containing data on project and harvest prices.
}
\description{
Download RMA projected and harvest prices used for revenue protection insurance plans.
}
\examples{
\dontrun{get_price_data(2024,"corn")}
\dontrun{get_price_data(year = 2022:2024, state = "VA")}
}
