% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfcip-package.R
\docType{package}
\name{rfcip-package}
\alias{rfcip}
\alias{rfcip-package}
\title{rfcip: Tools for accessing crop insurance data}
\description{
This package provides a set of tools to access public data on the Federal Crop Insurance Program, including both Summary of Business data and Actuarial Data Master (ADM) datasets.
}
\author{
\strong{Maintainer}: Dylan Turner \email{dylan.turner3@outlook.com} (\href{https://orcid.org/0000-0002-0915-7384}{ORCID})

}
\keyword{internal}
